+++ src/state/index.ts
@@ -1,1 +1,3 @@
 export * from './caseStore';
+export * from './casesStore';
+export * from './clientsStore';